+++ src/wr_ts.c
@@ -31,14 +31,13 @@ struct jls_wr_ts_s {
     uint16_t signal_id;
     enum jls_track_type_e track_type;
     uint32_t decimate_factor;
-    struct jls_index_s * index[JLS_SUMMARY_LEVEL_COUNT - 1];
-    struct jls_payload_header_s * summary[JLS_SUMMARY_LEVEL_COUNT - 1];
+    struct jls_index_s * index[JLS_SUMMARY_LEVEL_COUNT];                // level 0 not used
+    struct jls_payload_header_s * summary[JLS_SUMMARY_LEVEL_COUNT];     // level 0 not used
 };
 
-#if 0
 static void ts_free(struct jls_wr_ts_s * self) {
     if (self) {
-        for (int i = 0; i < JLS_SUMMARY_LEVEL_COUNT - 1; ++i) {
+        for (int i = 0; i < JLS_SUMMARY_LEVEL_COUNT; ++i) {
             if (self->index[i]) {
                 free(self->index[i]);
                 self->index[i] = 0;
@@ -68,7 +67,7 @@ static int32_t index_alloc(struct jls_wr_ts_s * self, uint8_t level) {
     index->header.entry_count = 0;
     index->header.entry_size_bits = sizeof(struct jls_index_entry_s) * 8;
     index->header.rsv16 = 0;
-    self->index[level - 1] = index;
+    self->index[level] = index;
     return 0;
 }
 
@@ -104,11 +103,15 @@ static int32_t summary_alloc(struct jls_wr_ts_s * self, uint8_t level) {
     summary->entry_count = 0;
     summary->entry_size_bits = (uint16_t) (8 * entry_sz);
     summary->rsv16 = 0;
-    self->summary[level - 1] = summary;
+    self->summary[level] = summary;
     return 0;
 }
-#endif
 
+static int32_t alloc(struct jls_wr_ts_s * self, uint8_t level) {
+    ROE(index_alloc(self, level));
+    ROE(summary_alloc(self, level));
+    return 0;
+}
 
 int32_t jls_wr_ts_open(
         struct jls_wr_ts_s ** instance,
@@ -128,41 +131,124 @@ int32_t jls_wr_ts_open(
     return 0;
 }
 
-static void ts_close(struct jls_wr_ts_s * self, uint8_t level) {
-    (void) self;
-    (void) level;
-    // todo
+static int32_t commit(struct jls_wr_ts_s * self, int level) {
+    if ((level < 1) || (level > JLS_SUMMARY_LEVEL_COUNT)) {
+        JLS_LOGE("invalid level");
+        return JLS_ERROR_PARAMETER_INVALID;
+    }
+    ROE(alloc(self, level + 1));
+
+    struct jls_index_s * index = self->index[level];
+    struct jls_payload_header_s * summary_header = self->summary[level];
+    if (!index || !summary_header) {
+        return JLS_ERROR_PARAMETER_INVALID;
+    }
+    if (!index->header.entry_count) {
+        return 0;
+    }
+
+    // update headers
+    index->header.timestamp = index->entries[0].timestamp;
+    summary_header->timestamp = index->entries[0].timestamp;
+
+    // write index
+    uint8_t * p_end = (uint8_t *) &index->entries[index->header.entry_count];
+    uint8_t * p_start = (uint8_t *) index;
+    uint32_t len = p_end - p_start;
+    ROE(jls_wr_index_prv(self->wr, self->signal_id, self->track_type, level, p_start, len));
+
+    // add to upper level and compute summary write
+    struct jls_index_s * index_up = self->index[level + 1];
+    struct jls_payload_header_s * summary_header_up = self->summary[level + 1];
+    index_up->entries[index_up->header.entry_count++] = index->entries[0];
+    if (self->track_type == JLS_TRACK_TYPE_ANNOTATION) {
+        struct jls_annotation_summary_s * summary = (struct jls_annotation_summary_s *) summary_header;
+        p_end = (uint8_t *) &summary->entries[summary->header.entry_count];
+        p_start = (uint8_t *) summary;
+        struct jls_annotation_summary_s * summary_up = (struct jls_annotation_summary_s *) summary_header_up;
+        summary_up->entries[summary_up->header.entry_count++] = summary->entries[0];
+    } else if (self->track_type == JLS_TRACK_TYPE_UTC) {
+        struct jls_utc_summary_s * summary = (struct jls_utc_summary_s *) summary_header;
+        p_end = (uint8_t *) &summary->entries[summary->header.entry_count];
+        p_start = (uint8_t *) summary;
+        struct jls_utc_summary_s * summary_up = (struct jls_utc_summary_s *) summary_header_up;
+        summary_up->entries[summary_up->header.entry_count++] = summary->entries[0];
+    }
+
+    // write summary.
+    len = p_end - p_start;
+    ROE(jls_wr_summary_prv(self->wr, self->signal_id, self->track_type, level, p_start, len));
+
+    // When up is full, commit it
+    if (index_up->header.entry_count >= self->decimate_factor) {
+        ROE(commit(self, level + 1));
+    }
+
+    // Reset our entry count since all have been written.
+    index->header.entry_count = 0;
+    summary_header->entry_count = 0;
+    return 0;
 }
 
-int32_t jls_wf_ts_close(struct jls_wr_ts_s * self) {
+int32_t jls_wr_ts_close(struct jls_wr_ts_s * self) {
     if (self) {
         for (uint8_t level = 1; level < JLS_SUMMARY_LEVEL_COUNT; ++level) {
-            ts_close(self, level);
+            commit(self, level);
         }
-        free(self);
+        ts_free(self);
     }
     return 0;
 }
 
 int32_t jls_wr_ts_anno(struct jls_wr_ts_s * self, int64_t timestamp, int64_t offset,
                        enum jls_annotation_type_e annotation_type, uint8_t group_id, float y) {
-    (void) self;
-    (void) timestamp;
-    (void) offset;
-    (void) annotation_type;
-    (void) group_id;
-    (void) y;
-    // todo add index entry
-    // todo add summary entry
+    if (self->track_type != JLS_TRACK_TYPE_ANNOTATION) {
+        JLS_LOGE("track_type mismatch");
+        return JLS_ERROR_PARAMETER_INVALID;
+    }
+    ROE(alloc(self, 1));
+    struct jls_index_s * index = self->index[1];
+    struct jls_annotation_summary_s * summary = (struct jls_annotation_summary_s *) self->summary[1];
+
+    struct jls_index_entry_s * index_entry = &index->entries[index->header.entry_count++];
+    index_entry->timestamp = timestamp;
+    index_entry->offset = offset;
+
+    struct jls_annotation_summary_entry_s * summary_entry = &summary->entries[summary->header.entry_count++];
+    summary_entry->timestamp = timestamp;
+    summary_entry->annotation_type = annotation_type;
+    summary_entry->group_id = group_id;
+    summary_entry->rsv8_1 = 0;
+    summary_entry->rsv8_2 = 0;
+    summary_entry->y = y;
+
+    if (index->header.entry_count >= self->decimate_factor) {
+        ROE(commit(self, 1));
+    }
+
     return 0;
 }
 
 int32_t jls_wr_ts_utc(struct jls_wr_ts_s * self, int64_t sample_id, int64_t offset, int64_t utc) {
-    (void) self;
-    (void) sample_id;
-    (void) offset;
-    (void) utc;
-    // todo add index entry
-    // todo add summary entry
+    if (self->track_type != JLS_TRACK_TYPE_UTC) {
+        JLS_LOGE("track_type mismatch");
+        return JLS_ERROR_PARAMETER_INVALID;
+    }
+    ROE(alloc(self, 1));
+    struct jls_index_s * index = self->index[1];
+    struct jls_utc_summary_s * summary = (struct jls_utc_summary_s *) self->summary[1];
+
+    struct jls_index_entry_s * index_entry = &index->entries[index->header.entry_count++];
+    index_entry->timestamp = sample_id;
+    index_entry->offset = offset;
+
+    struct jls_utc_summary_entry_s * summary_entry = &summary->entries[summary->header.entry_count++];
+    summary_entry->sample_id = sample_id;
+    summary_entry->timestamp = utc;
+
+    if (index->header.entry_count >= self->decimate_factor) {
+        ROE(commit(self, 1));
+    }
+
     return 0;
 }